requirements.txt
anthropic>=0.40.0
python-telegram-bot>=21.0
feedparser>=6.0.11
websockets>=12.0
pyyaml>=6.0.1
requests>=2.31.0
schedule>=1.2.1
pytest>=7.4.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
